requirements.txt
Django==1.10
Markdown==2.6.6
Pillow==3.3.0
argparse==1.2.1
django-filter==0.13.0
gunicorn==19.6.0
httpagentparser==1.7.8
redis==2.10.5
requests==2.10.0
wsgiref==0.1.2
